some smaller changes in light signal systems data formats and writer for the xml files
git-svn-id: file:///srv/local/data/cleanRepo/trunk@3233 a0517364-e873-4167-b1c8-416daa5a6b91

+++ matsim/src/org/matsim/basic/lightsignalsystems/BasicLane.java
@@ -32,7 +32,7 @@ import org.matsim.basic.v01.Id;
 public class BasicLane {
 
 	private Id id;
-	private double numberOfRepresentedLanes;
+	private int numberOfRepresentedLanes;
 	private double length;
 	private List<Id> toLinkIds;
 
@@ -46,7 +46,7 @@ public class BasicLane {
 	/**
 	 * @param number
 	 */
-	public void setNumberOfRepresentedLanes(double number) {
+	public void setNumberOfRepresentedLanes(int number) {
 		this.numberOfRepresentedLanes = number;
 	}
 
@@ -59,7 +59,7 @@ public class BasicLane {
 	}
 
 	
-	public double getNumberOfRepresentedLanes() {
+	public int getNumberOfRepresentedLanes() {
 		return numberOfRepresentedLanes;
 	}
